soda 2.0

+++ Items/SalvagedCrate.cs
@@ -67,8 +67,8 @@ namespace SGAmod.Items
 					if (Main.netMode > 0)
 					{
 						ModPacket packet = mod.GetPacket();
-							packet.Write((int)75);
-							packet.Write(mod.NPCType("CratrosityPML"));
+						packet.Write((ushort)75);
+						packet.Write(mod.NPCType("CratrosityPML"));
 							packet.Write((int)(ply.Center.X	- 1600));
 							packet.Write(3200);
 							packet.Write(ply.whoAmI);
